// ApartmentApps.Client.Mobile/src/services/backend/generated/backend.generated.ts
export const API_BASE_URL = new OpaqueToken('API_BASE_URL');
// ...
export class SearchEnginesClient {
    private http: Http = null; 
    private baseUrl: string = undefined; 
    protected jsonParseReviver: (key: string, value: any) => any = undefined;

    constructor(@Inject(Http) http: Http, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http; 
        this.baseUrl = baseUrl ? baseUrl : "http://localhost:54686"; 
    }

    /**
     * @return OK
     */
    getSearchModel(id: string): Observable<SearchModelGetResponse> {
        let url_ = this.baseUrl + "/api/searches?"; 
        if (id === undefined || id === null)
            throw new Error("The parameter 'id' must be defined and cannot be null.");
        else
            url_ += "id=" + encodeURIComponent("" + id) + "&";

        const content_ = "";
        
        return this.http.request(url_, {
            body: content_,
            method: "get",
            headers: new Headers({
                "Content-Type": "application/json; charset=UTF-8"
            })
        }).map((response) => {
            return this.processGetSearchModel(response);
        }).catch((response: any, caught: any) => {
            if (response instanceof Response) {
                try {
                    return Observable.of(this.processGetSearchModel(response));
                } catch (e) {
                    return <Observable<SearchModelGetResponse>><any>Observable.throw(e);
                }
            } else
                return <Observable<SearchModelGetResponse>><any>Observable.throw(response);
        });
    }

    protected processGetSearchModel(response: Response): SearchModelGetResponse {
        const responseText = response.text();
        const status = response.status; 

		if (status === 200) {
            let result200: SearchModelGetResponse = null;
            let resultData200 = responseText === "" ? null : JSON.parse(responseText, this.jsonParseReviver);
            result200 = resultData200 ? SearchModelGetResponse.fromJS(resultData200) : new SearchModelGetResponse();
            return result200;
        } else if (status !== 200 && status !== 204) {
            this.throwException("An unexpected server error occurred.", status, responseText);
        }
        return null;
    }

    protected throwException(message: string, status: number, response: string, result?: any): any {
    	if(result !== null && result !== undefined)
    		throw result;
    	else
    		throw new SwaggerException(message, status, response);
    }
}

@Injectable()
export class VersionClient {
    private http: Http = null; 
    private baseUrl: string = undefined; 
    protected jsonParseReviver: (key: string, value: any) => any = undefined;

    constructor(@Inject(Http) http: Http, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http; 
        this.baseUrl = baseUrl ? baseUrl : "http://localhost:54686"; 
    }

    /**
     * @return OK
     */
    get(): Observable<VersionInfo> {
        let url_ = this.baseUrl + "/api/Version";

        const content_ = "";
        
        return this.http.request(url_, {
            body: content_,
            method: "get",
            headers: new Headers({
                "Content-Type": "application/json; charset=UTF-8"
            })
        }).map((response) => {
            return this.processGet(response);
        }).catch((response: any, caught: any) => {
            if (response instanceof Response) {
                try {
                    return Observable.of(this.processGet(response));
                } catch (e) {
                    return <Observable<VersionInfo>><any>Observable.throw(e);
                }
            } else
                return <Observable<VersionInfo>><any>Observable.throw(response);
        });
    }

    protected processGet(response: Response): VersionInfo {
        const responseText = response.text();
        const status = response.status; 

		if (status === 200) {
            let result200: VersionInfo = null;
            let resultData200 = responseText === "" ? null : JSON.parse(responseText, this.jsonParseReviver);
            result200 = resultData200 ? VersionInfo.fromJS(resultData200) : new VersionInfo();
            return result200;
        } else if (status !== 200 && status !== 204) {
            this.throwException("An unexpected server error occurred.", status, responseText);
        }
        return null;
    }

    protected throwException(message: string, status: number, response: string, result?: any): any {
    	if(result !== null && result !== undefined)
    		throw result;
    	else
    		throw new SwaggerException(message, status, response);
    }
}
// ...
export class QueryResultOfLookupBindingModel { 
    total: number; 
    result: LookupBindingModel[];

    constructor(data?: any) {
        if (data !== undefined) {
            this.total = data["Total"] !== undefined ? data["Total"] : null;
            if (data["Result"] && data["Result"].constructor === Array) {
                this.result = [];
                for (let item of data["Result"])
                    this.result.push(LookupBindingModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): QueryResultOfLookupBindingModel {
        return new QueryResultOfLookupBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["Total"] = this.total !== undefined ? this.total : null;
        if (this.result && this.result.constructor === Array) {
            data["Result"] = [];
            for (let item of this.result)
                data["Result"].push(item.toJS());
        }
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new QueryResultOfLookupBindingModel(JSON.parse(json));
    }
}

export class LookupBindingModel { 
    id: string; 
    title: string; 
    textPrimary: string; 
    textSecondary: string; 
    imageUrl: string; 
    selected: boolean;

    constructor(data?: any) {
        if (data !== undefined) {
            this.id = data["Id"] !== undefined ? data["Id"] : null;
            this.title = data["Title"] !== undefined ? data["Title"] : null;
            this.textPrimary = data["TextPrimary"] !== undefined ? data["TextPrimary"] : null;
            this.textSecondary = data["TextSecondary"] !== undefined ? data["TextSecondary"] : null;
            this.imageUrl = data["ImageUrl"] !== undefined ? data["ImageUrl"] : null;
            this.selected = data["Selected"] !== undefined ? data["Selected"] : null;
        }
    }

    static fromJS(data: any): LookupBindingModel {
        return new LookupBindingModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["Id"] = this.id !== undefined ? this.id : null;
        data["Title"] = this.title !== undefined ? this.title : null;
        data["TextPrimary"] = this.textPrimary !== undefined ? this.textPrimary : null;
        data["TextSecondary"] = this.textSecondary !== undefined ? this.textSecondary : null;
        data["ImageUrl"] = this.imageUrl !== undefined ? this.imageUrl : null;
        data["Selected"] = this.selected !== undefined ? this.selected : null;
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new LookupBindingModel(JSON.parse(json));
    }
}
// ...
export class SearchModelGetResponse { 
    model: ClientSearchModel;

    constructor(data?: any) {
        if (data !== undefined) {
            this.model = data["Model"] ? ClientSearchModel.fromJS(data["Model"]) : null;
        }
    }

    static fromJS(data: any): SearchModelGetResponse {
        return new SearchModelGetResponse(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["Model"] = this.model ? this.model.toJS() : null;
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new SearchModelGetResponse(JSON.parse(json));
    }
}

export class ClientSearchModel { 
    id: string; 
    filters: ClientSearchFilterModel[];

    constructor(data?: any) {
        if (data !== undefined) {
            this.id = data["Id"] !== undefined ? data["Id"] : null;
            if (data["Filters"] && data["Filters"].constructor === Array) {
                this.filters = [];
                for (let item of data["Filters"])
                    this.filters.push(ClientSearchFilterModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): ClientSearchModel {
        return new ClientSearchModel(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["Id"] = this.id !== undefined ? this.id : null;
        if (this.filters && this.filters.constructor === Array) {
            data["Filters"] = [];
            for (let item of this.filters)
                data["Filters"].push(item.toJS());
        }
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new ClientSearchModel(JSON.parse(json));
    }
}
// ...
export class VersionInfo { 
    androidStoreUrl: string; 
    iPhoneStoreUrl: string; 
    version: number; 
    iPhoneBuildNumber: number; 
    androidBuildNumber: number;

    constructor(data?: any) {
        if (data !== undefined) {
            this.androidStoreUrl = data["AndroidStoreUrl"] !== undefined ? data["AndroidStoreUrl"] : null;
            this.iPhoneStoreUrl = data["IPhoneStoreUrl"] !== undefined ? data["IPhoneStoreUrl"] : null;
            this.version = data["Version"] !== undefined ? data["Version"] : null;
            this.iPhoneBuildNumber = data["IPhoneBuildNumber"] !== undefined ? data["IPhoneBuildNumber"] : null;
            this.androidBuildNumber = data["AndroidBuildNumber"] !== undefined ? data["AndroidBuildNumber"] : null;
        }
    }

    static fromJS(data: any): VersionInfo {
        return new VersionInfo(data);
    }

    toJS(data?: any) {
        data = data === undefined ? {} : data;
        data["AndroidStoreUrl"] = this.androidStoreUrl !== undefined ? this.androidStoreUrl : null;
        data["IPhoneStoreUrl"] = this.iPhoneStoreUrl !== undefined ? this.iPhoneStoreUrl : null;
        data["Version"] = this.version !== undefined ? this.version : null;
        data["IPhoneBuildNumber"] = this.iPhoneBuildNumber !== undefined ? this.iPhoneBuildNumber : null;
        data["AndroidBuildNumber"] = this.androidBuildNumber !== undefined ? this.androidBuildNumber : null;
        return data; 
    }

    toJSON() {
        return JSON.stringify(this.toJS());
    }

    clone() {
        const json = this.toJSON();
        return new VersionInfo(JSON.parse(json));
    }
}
// ...
export class SwaggerException extends Error {
    message: string;
    status: number; 
    response: string; 
    result?: any; 

    constructor(message: string, status: number, response: string, result?: any) {
        super();

        this.message = message;
        this.status = status;
        this.response = response;
        this.result = result;
    }
}